pgvector/pgvector-openshift/utils/requirements.txt
pgvector==0.4.1
psycopg2-binary==2.9.10
requests==2.32.4
